+++ apps/frontend-manage/src/components/courses/LiveQuizList.tsx
@@ -28,6 +28,8 @@ function LiveQuizList({
     <div className="flex w-full flex-col items-end">
       <div className="flex flex-row gap-2">
         <QRCodePopover
+          triggerStyle="basic"
+          triggerText={t('manage.general.qrCode')}
           relHref={`/course/${courseId}/liveQuizzes`}
           data={{ cy: `qr-link-live-quiz-list` }}
         />