+++ packages/graphql/src/services/stacks.ts
@@ -3177,6 +3177,7 @@ export async function respondToElementStack(
     })
 
     if (
+      !isOwner &&
       stack?.microLearning &&
       (stack.elements.some((element) => element.responses.length > 0) ||
         dayjs().isAfter(dayjs(stack.microLearning.scheduledEndAt)))